Mover ArgsResolver to a separate file
